tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/065.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: cloneNode with external script, changed .src</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
	<script src="scripts/include-1.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		var script=document.getElementsByTagName('script')[3].cloneNode(true);
		script.src='scripts/include-2.js'
		document.body.appendChild( script );
		log('end script #1');
	</script>
	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['external script #1', 'inline script #1', 'end script #1', 'inline script #2']);
		t.done();
}
	onload = t.step_func(test);
	</script>

</body></html>
